# openapi.yaml
openapi: 3.1.0
info:
  title: BetterReads API
  description: Book tracking and review API.
  contact:
    name: BetterReads
  license:
    name: Apache 2.0
    url: https://github.com/dahl-jar/betterreads/blob/main/LICENSE
  version: v1
servers:
- url: http://localhost:8080
  description: Production
security:
- bearer-jwt: []
...
paths:
  /api/v1/me/books/{key}/status:
    put:
      tags:
      - Shelf
      summary: Set the shelf status for a book
      operationId: changeStatus
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetStatusRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: The updated shelf entry
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ShelfEntryResponse"
        "400":
          description: Unknown status value
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: No book with that key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/me/books/{key}/favorite:
    put:
      tags:
      - Shelf
      summary: Set the favorite flag for a book
      operationId: markFavorite
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetFavoriteRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: The updated shelf entry
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ShelfEntryResponse"
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: No book with that key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Rotate access and refresh tokens
      operationId: refresh
      parameters:
      - name: br_refresh
        in: cookie
        required: false
        schema:
          type: string
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: "New access token, refresh cookie rotated"
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/AuthResponse"
        "401":
          description: "Missing, expired, or already-rotated refresh token"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
...
  /api/v1/me/books/{key}:
    delete:
      tags:
      - Shelf
      summary: Remove a book from the shelf
      operationId: remove_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "503":
          description: Service Unavailable
        "204":
          description: Book removed or already absent
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: No book with that key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    patch:
      tags:
      - Shelf
      summary: Update reading dates and note
      operationId: update
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntryRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: The updated shelf entry
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ShelfEntryResponse"
        "400":
          description: Finished date before started date
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Book is not on the shelf
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/search/books:
    get:
      tags:
      - Search
      summary: Search the book catalog
      operationId: search
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          maxLength: 200
          minLength: 0
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookSearchDocument"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
      security: []
  /api/v1/search/books/events:
    get:
      tags:
      - Search
      summary: Stream newly added books that match the query
      operationId: streamHits
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          maxLength: 200
          minLength: 0
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: A search-hit event for each newly added book matching q.
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/BookSearchDocument"
      security: []
...
  /api/v1/me/books:
    get:
      tags:
      - Shelf
      summary: List the current user's shelf
      operationId: list
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - WANT_TO_READ
          - CURRENTLY_READING
          - FINISHED
          - DROPPED
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: "The shelf, newest first"
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShelfEntryResponse"
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/books:
    get:
      tags:
      - Catalog
      summary: Get a homepage book list
      operationId: list_1
      parameters:
      - name: list
        in: query
        required: true
        schema:
          type: string
          enum:
          - RECENTLY_ADDED
          - TOP_RATED
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 12
          maximum: 50
          minimum: 1
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookCardResponse"
      security: []
...
components:
  schemas:
    SetStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum:
          - WANT_TO_READ
          - CURRENTLY_READING
          - FINISHED
          - DROPPED
      required:
      - status
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
    ShelfEntryResponse:
      type: object
      properties:
        key:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        coverUrl:
          type: string
        status:
          type: string
          enum:
          - WANT_TO_READ
          - CURRENTLY_READING
          - FINISHED
          - DROPPED
        favorite:
          type: boolean
        startedAt:
          type: string
          format: date
        finishedAt:
          type: string
          format: date
        notes:
          type: string
        addedAt:
          type: string
          format: date
        averageRating:
          type: number
        myRating:
          type: integer
          format: int32
    SetFavoriteRequest:
      type: object
      properties:
        favorite:
          type: boolean
      required:
      - favorite
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: Short-lived access JWT
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzMwMDAwMDAwfQ.signature
        user:
          $ref: "#/components/schemas/UserResponse"
    UserResponse:
      type: object
      properties:
        username:
          type: string
          example: john.doe
        email:
          type: string
          example: john.doe@example.com
        emailVerified:
          type: boolean
          example: true
        displayName:
          type: string
          example: John Doe
        avatarUrl:
          type: string
          example: https://cdn.example.com/u/john-doe.png
        bio:
          type: string
          example: Reads mostly sci-fi.
...
    UpdateEntryRequest:
      type: object
      properties:
        startedAt:
          type: string
          format: date
        finishedAt:
          type: string
          format: date
        notes:
          type: string
          maxLength: 2000
          minLength: 0
...
    BookSearchDocument:
      type: object
      properties:
        bookId:
          type: string
        title:
          type: string
        subtitle:
          type: string
        seriesName:
          type: string
        seriesPosition:
          type: integer
          format: int32
        authors:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        language:
          type: string
        coverUrl:
          type: string
        publicationYear:
          type: integer
          format: int32
        popularityScore:
          type: number
          format: double
...
    BookCardResponse:
      type: object
      properties:
        key:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        coverUrl:
          type: string
        firstPublishYear:
          type: integer
          format: int32
        averageRating:
          type: number
        ratingCount:
          type: integer
          format: int32
...
    ResponseMeta:
      properties:
        total:
          type: integer
          format: int64
          example: 142
        offset:
          type: integer
          format: int32
          example: 0
        limit:
          type: integer
          format: int32
          example: 20
  securitySchemes:
    bearer-jwt:
      type: http
      description: Access JWT from POST /api/v1/auth/login.
      scheme: bearer
      bearerFormat: JWT